Riddhima
As a Project Management Intern at Genpact, I work with the PMO team of the Finance Department to research and communicate the value of the PMO services to potential clients. I contribute to the creation of research reports and presentations, and help in organizing and visualizing data to convey meaningful insights to stakeholders. I also track and generate project documents for clients, and analyze compliance and non-compliance percentages of different verticals.

I am an undergraduate economics student at Symbiosis School of Economics, with a strong interest in Finance, Business Development, and Marketing. I aim to build a start-up that leverages my quantitative and qualitative research skills, as well as my knowledge of statistics, power BI, and data science. I have completed multiple online courses to enhance my skills and knowledge in these domains. I am also active in various social pursuits and enjoy exercising and listening to music in my spare time.